Explain what Pelican is on home page
The problem with only showing recent posts on getpelican.com is that
it's not clear to visitors what Pelican is or what it does. This
replaces the list of recent posts on the home page with a static page
describing the Pelican project. Recent posts are still accessible via
the News category.